minor up-date
