+++ test_productHandler.py
@@ -1,7 +1,6 @@
 # Imports
 from phangsPipeline import keyHandler as kh
-#from phangsPipeline import productHandler as prh
-from phangsPipeline import productHandler_dzliu_20200226a as prh
+from phangsPipeline import productHandler as prh
 
 
 # Instantiate handlers
